Crate arm_gic

ARM Generic Interrupt Controller (GIC) register definitions and basic operations. Driver for the Arm Generic Interrupt Controller version 2 or 3 or 4, on aarch64.

This top level module contains functions that are not specific to any particular interrupt controller, as support for other GIC versions may be added in future.

Note:

Interrupt grouping(secure state) is not supported

Interrupt proiority(preempt) is not supported

Please contact the developer if you need this function

Structs

GicV2

Driver for an Arm Generic Interrupt Controller version 2.

GicV3

Driver for an Arm Generic Interrupt Controller version 3 (or 4).

IntId

An interrupt ID.

Enums

InterruptType

Different types of interrupt that the GIC handles.

TriggerMode

Interrupt trigger mode.

Traits

GenericArmGic

GenericArmGic. It is used to implement the interface abstraction that the interrupt chip driver should provide to the outside world. I hope that the versatility of the driver interface should support more chip architectures.

Functions

translate_irq

Translate an interrupt of a given type to a GIC INTID.